Refactor selection to a selection service
